\name{rc.get.params}
\docType{package}
\alias{rc.get.params}
\title{Get Parameters}
\description{
  Retrieve parameters of Circos plot.
}
\usage{rc.get.params()}
\value{
 A list.
}
\author{
Minghui Wang <m.h.wang@live.com>
}
\seealso{\code{\link{rc.initialize}}, \code{\link{rc.reset.params}}}
\examples{
##not run###
}
