\name{rc.reset.params}
\docType{package}
\alias{rc.reset.params}
\title{Reset Circos Plot Parameters}
\description{
  Reset parameters for Circos plot.
}
\usage{rc.reset.params(params)}
\arguments{
 \item{params}{list.}
}
\author{
Minghui Wang <m.h.wang@live.com>
}
\seealso{\code{\link{rc.initialize}}, \code{\link{rc.get.params}}}
\examples{
##not run###
}
