\name{rc.reset.params}
\alias{rc.reset.params}
\title{Reset Circos Plot Parameters}
\description{
  Reset parameters for Circos plot.
}
\usage{rc.reset.params(params)}
\arguments{
 \item{params}{list.}
}
\author{
Minghui Wang <m.h.wang@live.com>
}
\seealso{\code{\link{rc.initialize}}, \code{\link{rc.get.params}}}
\examples{
#This is not to be run alone. Please see tutorial vignette("netweaver") for usage.
}
