\name{rc.track.pos}
\docType{package}
\alias{rc.track.pos}
\title{Get Track Positions}
\description{
  Calculate inner and outer radius of a track.
}
\usage{rc.track.pos(track.id)}
\arguments{
 \item{track.id}{integer, number of the track from the outermost.}
}
\value{
 A vector with two radius values: out.pos and in.pos.
}
\author{
Minghui Wang <m.h.wang@live.com>
}
\seealso{\code{\link{rc.initialize}}, \code{\link{rc.reset.params}}}
\examples{
#This is not to be run alone. Please see tutorial vignette("netweaver") for usage.
}
